// packages/components/src/ui/Tag/Tag.styles.d.ts
export declare const useStyles: () => {
  tagContainer: {
    display: 'block';
  };
  tag: {
    borderRadius: import('packages/theme/src/constants/Radius').Radius;
  };
  tagDefault: {
    color: '#151617';
    backgroundColor: '#EFF1F6';
  };
  tagAction: {
    color: '#151617';
    backgroundColor: '#D2E8FE';
  };
  tagSm: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    paddingTop: number;
    paddingBottom: number;
  };
  tagMd: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    paddingTop: import('@alveole/theme').Spacing;
    paddingBottom: import('@alveole/theme').Spacing;
  };
};
//# sourceMappingURL=Tag.styles.d.ts.map